chore: rename spinner styles file
