+++ src/styles/index.ts
@@ -1,2 +1,2 @@
-export * from './layout-protected.styles';
+export * from './loading-spinner.styles';
 export * from './content-wrapper.styles';
